requirements.txt
# Core LLM and embeddings
openai>=1.6.0
langchain>=1.2.0
langchain-openai>=1.1.6
langchain-community>=0.4.1

# Vector database
chromadb>=1.4.0

# Document loaders
pypdf>=3.17.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
